Search for an existing member by mobile or phone
NationBuilder /people/add upsert endpoint will create a duplicate
if the phone matches an existing user but the email passed is different.

To avoid this, first lookup the user by mobile if exists, or fallback
to phone. If the member exists, use the member (and do not update
anything). Otherwse, fallback to using the upsert endpoint.